jellyfish-address refactor (#238)
* updated order of jellyfish-address

* rename file to lower snakecase

* rename test class

* fix and rename

* reorder packages in introduction.md
