+++ .github/labeler.yml
@@ -16,6 +16,11 @@ labels:
     matcher:
       files: "packages/jellyfish/**"
 
+  - label: area/jellyfish-address
+    sync: true
+    matcher:
+      files: "packages/jellyfish-address/**"
+
   - label: area/jellyfish-api-core
     sync: true
     matcher:
@@ -50,12 +55,6 @@ labels:
     sync: true
     matcher:
       files: "packages/jellyfish-wallet-mnemonic/**"
-
-
-  - label: area/jellyfish-address
-    sync: true
-    matcher:
-      files: "packages/jellyfish-address/**"
 
   - label: area/testcontainers
     sync: true
