+++ .github/labels.yml
@@ -59,6 +59,8 @@
 - color: fbca04
   name: area/jellyfish
 - color: fbca04
+  name: area/jellyfish-address
+- color: fbca04
   name: area/jellyfish-api-core
 - color: fbca04
   name: area/jellyfish-api-jsonrpc
@@ -74,8 +76,6 @@
   name: area/jellyfish-wallet
 - color: fbca04
   name: area/jellyfish-wallet-mnemonic
-- color: fbca04
-  name: area/jellyfish-address
 - color: fbca04
   name: area/testcontainers
 - color: fbca04
