+++ packages/jellyfish-address/jest.config.js
@@ -1,7 +1,6 @@
 module.exports = {
   testEnvironment: 'node',
   testMatch: [
-    '**/__testcontainers__/**/*.test.ts',
     '**/__tests__/**/*.test.ts'
   ],
   transform: {
