+++ packages/jellyfish-address/src/index.ts
@@ -1,20 +1,20 @@
 
 import { getNetwork, NetworkName } from '@defichain/jellyfish-network'
-import { Address, AddressType, UnknownTypeAddress } from './Address'
-import { Base58Address } from './Base58Address'
-import { Bech32Address } from './Bech32Address'
-import { P2PKH } from './P2PKH'
-import { P2SH } from './P2SH'
-import { P2WSH } from './P2WSH'
-import { P2WPKH } from './P2WPKH'
+import { Address, AddressType, UnknownTypeAddress } from './address'
+import { Base58Address } from './base58_address'
+import { Bech32Address } from './bech32_address'
+import { P2PKH } from './p2pkh'
+import { P2SH } from './p2sh'
+import { P2WSH } from './p2wsh'
+import { P2WPKH } from './p2wpkh'
 
-export * from './Address'
-export * from './Base58Address'
-export * from './Bech32Address'
-export * from './P2PKH'
-export * from './P2SH'
-export * from './P2WPKH'
-export * from './P2WSH'
+export * from './address'
+export * from './base58_address'
+export * from './bech32_address'
+export * from './p2pkh'
+export * from './p2sh'
+export * from './p2wpkh'
+export * from './p2wsh'
 
 /**
  * When insist to use the "network" decoded from raw address, instead of passing one based on running application environment
